Convert font to a custom resource with fixes
Improve notifications by using converge_by, simplify checking the current state, and rename the file property to name property since file causes errors in modern Chef
